Keep clearing out Supabase
