+++ package.json
@@ -17,9 +17,7 @@
   "dependencies": {
     "@nuxt/image": "^1.8.1",
     "@nuxt/ui": "^2.18.7",
-    "@nuxtjs/supabase": "^1.4.1",
     "@pinia/nuxt": "^0.5.5",
-    "@supabase/gotrue-js": "^2.65.0",
     "nuxt": "^3.13.2",
     "pinia": "^2.2.4",
     "vue": "latest",
